feat(streaming): M6c — TRIMA streams in C, completing the M6 duals tranche (150 funcs)
TRIMA's batch is `if(period is odd){odd-weight rings}else{even-weight rings}` —
two genuinely different steady loops over triangular-weight rings, selected by
period parity. The stream keeps the batch untouched and transcribes BOTH arms as
inline dual-mode steady loops sharing one handle; the mode is fixed at Open from
the immutable stored optInTimePeriod and Step branches on it (no mode tag — the
Dispatch precedent).

This is the if/else form of the M6a dual-mode machinery (DI/DM were the
early-return form) and the first dual-mode with non-scalar (ring) state:
emit_dual_state_struct asserts both arms' ring layout is byte-identical before
unioning the per-arm scalars into one struct.

Flagging TRIMA `stream` also completes the MA-dispatch join: every single-output
MA now streams, so TRIMA becomes a supported optInMAType arm (the derivation
picks it up automatically from the callee stream flag). A new generator test,
every_matype_is_streamable_except_tracked_blockers, is the forever contract: it
fails if any MAType stops streaming (blocklist = [mama], must only shrink), so a
future non-streaming MAType can never silently break the MA-family streams' batch
bit-exactness — an un-streamable MAType would make every MA-parameterised
function non-bit-exact against its batch.

Also folds in a TRIMA UB fix surfaced by review: the triangular-weight `factor`
was computed in `int` and overflowed past period ~92682; widened to `double`
(bit-identical below the overflow threshold — fuzz-064 unchanged).

- streaming.rs: analyze_dual_mode's if/else form (both arms a steady loop,
  neither returns, condition not a `<=literal` threshold — that routes to
  FastPathSkip); DualModePlan gained an epilogue field.
- c_stream.rs: emit_dual_mode / emit_dual_state_struct handle non-scalar (ring)
  state — union the per-arm scalars, share the byte-identical rings, branch
  Step/Open on the stored period, drop each arm's unused locals.

Gates: 4-language codegen 150 stream funcs / 10878 legs bit-exact + 161x4 pass;
TRIMA bit-exact incl. odd period 71; fuzz-064 135686 comparisons 0 failures
(batch bit-identical to 0.6.4); sanitizer 150 funcs / 313 legs ASan/UBSan/LSan
clean; clippy 1.96/1.97 + cargo test (+TRIMA plan pins, +MAType contract) +
regen-oracle idempotent. Adversarial review: 0 blockers.

+++ java/src/com/tictactec/ta/lib/Core.java
@@ -59574,15 +59574,20 @@ public class Core {
  *  -------------------------------------------------------------------
  *  MF       Mario Fortier
  *  CR       Chris (crokusek@hotmail.com)
+ *  CC       Claude Code (AI assistant)
  *
  * Change history:
  *
- *  MMDDYY BY   Description
+ *  MMDDYY BY     Description
  *  -------------------------------------------------------------------
- *  010503 MF   Initial Coding
- *  031703 MF   Fix #701060. Correct logic when using a range with
- *              startIdx/endIdx. Thanks to Chris for reporting this.
- *  052603 MF   Adapt code to compile with .NET Managed C++
+ *  010503 MF     Initial Coding
+ *  031703 MF     Fix #701060. Correct logic when using a range with
+ *                startIdx/endIdx. Thanks to Chris for reporting this.
+ *  052603 MF     Adapt code to compile with .NET Managed C++
+ *  071226 MF,CC  Widen the triangular-weight factor to double: (i+1)*(i+1)
+ *                and i*(i+1) overflowed a 32-bit int at extreme periods
+ *                (past ~92682), silently returning garbage. Bit-identical
+ *                for every period where the int product fits.
  */
 
    public int trimaLookback( int optInTimePeriod )
@@ -59753,11 +59758,14 @@ public class Core {
           *    1+2+3+3+2+1 = n*(n+1)
           *                = 3 * 4 = 12
           */
-         /* Note: entirely done with int and becomes double only
-          *       on assignement to the factor variable.
+         /* Note: the (i+1) factors are widened to double so the product
+          *       cannot overflow a 32-bit int at extreme periods (i+1 reaches
+          *       ~50000 near the API maximum, and (i+1)*(i+1) exceeds INT_MAX
+          *       past period ~92682). For every period where the int product
+          *       fits, the widened value is identical.
           */
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          /* Initialize all the variable before
           * starting to iterate for each output.
@@ -59817,7 +59825,8 @@ public class Core {
           *  - Adjustment of numeratorAdd is different. See Step (2).
           */
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
+         /* widen: i*(i+1) overflows int past period ~92682 */
          factor = 1.0 / factor;
          /* Initialize all the variable before
           * starting to iterate for each output.
@@ -59903,7 +59912,7 @@ public class Core {
       outIdx = 0;
       if( optInTimePeriod % 2 == 1 ) {
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i;
@@ -59941,7 +59950,7 @@ public class Core {
          }
       } else {
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i - 1;
@@ -60024,7 +60033,7 @@ public class Core {
       outIdx = 0;
       if( optInTimePeriod % 2 == 1 ) {
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i;
@@ -60062,7 +60071,7 @@ public class Core {
          }
       } else {
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i - 1;
@@ -60134,7 +60143,7 @@ public class Core {
       outIdx = 0;
       if( optInTimePeriod % 2 == 1 ) {
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i;
@@ -60172,7 +60181,7 @@ public class Core {
          }
       } else {
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i - 1;
