+++ ta_codegen/input/trima/trima.c
@@ -4,15 +4,20 @@
  *  -------------------------------------------------------------------
  *  MF       Mario Fortier
  *  CR       Chris (crokusek@hotmail.com)
+ *  CC       Claude Code (AI assistant)
  *
  * Change history:
  *
- *  MMDDYY BY   Description
+ *  MMDDYY BY     Description
  *  -------------------------------------------------------------------
- *  010503 MF   Initial Coding
- *  031703 MF   Fix #701060. Correct logic when using a range with
- *              startIdx/endIdx. Thanks to Chris for reporting this.
- *  052603 MF   Adapt code to compile with .NET Managed C++
+ *  010503 MF     Initial Coding
+ *  031703 MF     Fix #701060. Correct logic when using a range with
+ *                startIdx/endIdx. Thanks to Chris for reporting this.
+ *  052603 MF     Adapt code to compile with .NET Managed C++
+ *  071226 MF,CC  Widen the triangular-weight factor to double: (i+1)*(i+1)
+ *                and i*(i+1) overflowed a 32-bit int at extreme periods
+ *                (past ~92682), silently returning garbage. Bit-identical
+ *                for every period where the int product fits.
  *
  */
 
@@ -172,11 +177,14 @@ TA_RetCode trima(int startIdx, int endIdx,
        *                = 3 * 4 = 12
        */
 
-      /* Note: entirely done with int and becomes double only
-       *       on assignement to the factor variable.
+      /* Note: the (i+1) factors are widened to double so the product
+       *       cannot overflow a 32-bit int at extreme periods (i+1 reaches
+       *       ~50000 near the API maximum, and (i+1)*(i+1) exceeds INT_MAX
+       *       past period ~92682). For every period where the int product
+       *       fits, the widened value is identical.
        */
       i = (optInTimePeriod>>1);
-      factor = (i+1)*(i+1);
+      factor = (double)(i+1)*(i+1);
       factor = 1.0/factor;
 
       /* Initialize all the variable before
@@ -248,7 +256,7 @@ TA_RetCode trima(int startIdx, int endIdx,
        *  - Adjustment of numeratorAdd is different. See Step (2).
        */
       i = (optInTimePeriod>>1);
-      factor = i*(i+1);
+      factor = (double)i*(i+1);   /* widen: i*(i+1) overflows int past period ~92682 */
       factor = 1.0/factor;
 
       /* Initialize all the variable before
