+++ ta_codegen/output/java/Core_TRIMA.java
@@ -4,15 +4,20 @@
  *  -------------------------------------------------------------------
  *  MF       Mario Fortier
  *  CR       Chris (crokusek@hotmail.com)
+ *  CC       Claude Code (AI assistant)
  *
  * Change history:
  *
- *  MMDDYY BY   Description
+ *  MMDDYY BY     Description
  *  -------------------------------------------------------------------
- *  010503 MF   Initial Coding
- *  031703 MF   Fix #701060. Correct logic when using a range with
- *              startIdx/endIdx. Thanks to Chris for reporting this.
- *  052603 MF   Adapt code to compile with .NET Managed C++
+ *  010503 MF     Initial Coding
+ *  031703 MF     Fix #701060. Correct logic when using a range with
+ *                startIdx/endIdx. Thanks to Chris for reporting this.
+ *  052603 MF     Adapt code to compile with .NET Managed C++
+ *  071226 MF,CC  Widen the triangular-weight factor to double: (i+1)*(i+1)
+ *                and i*(i+1) overflowed a 32-bit int at extreme periods
+ *                (past ~92682), silently returning garbage. Bit-identical
+ *                for every period where the int product fits.
  */
 
    public int trimaLookback( int optInTimePeriod )
@@ -183,11 +188,14 @@
           *    1+2+3+3+2+1 = n*(n+1)
           *                = 3 * 4 = 12
           */
-         /* Note: entirely done with int and becomes double only
-          *       on assignement to the factor variable.
+         /* Note: the (i+1) factors are widened to double so the product
+          *       cannot overflow a 32-bit int at extreme periods (i+1 reaches
+          *       ~50000 near the API maximum, and (i+1)*(i+1) exceeds INT_MAX
+          *       past period ~92682). For every period where the int product
+          *       fits, the widened value is identical.
           */
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          /* Initialize all the variable before
           * starting to iterate for each output.
@@ -247,7 +255,8 @@
           *  - Adjustment of numeratorAdd is different. See Step (2).
           */
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
+         /* widen: i*(i+1) overflows int past period ~92682 */
          factor = 1.0 / factor;
          /* Initialize all the variable before
           * starting to iterate for each output.
@@ -333,7 +342,7 @@
       outIdx = 0;
       if( optInTimePeriod % 2 == 1 ) {
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i;
@@ -371,7 +380,7 @@
          }
       } else {
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i - 1;
@@ -454,7 +463,7 @@
       outIdx = 0;
       if( optInTimePeriod % 2 == 1 ) {
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i;
@@ -492,7 +501,7 @@
          }
       } else {
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i - 1;
@@ -564,7 +573,7 @@
       outIdx = 0;
       if( optInTimePeriod % 2 == 1 ) {
          i = optInTimePeriod >> 1;
-         factor = (i + 1) * (i + 1);
+         factor = (double)(i + 1) * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i;
@@ -602,7 +611,7 @@
          }
       } else {
          i = optInTimePeriod >> 1;
-         factor = i * (i + 1);
+         factor = (double)i * (i + 1);
          factor = 1.0 / factor;
          trailingIdx = startIdx - lookbackTotal;
          middleIdx = trailingIdx + i - 1;
