+++ ta_codegen/output/rust/src/ta_func/trima.rs
@@ -45,15 +45,20 @@
  *  -------------------------------------------------------------------
  *  MF       Mario Fortier
  *  CR       Chris (crokusek@hotmail.com)
+ *  CC       Claude Code (AI assistant)
  *
  * Change history:
  *
- *  MMDDYY BY   Description
+ *  MMDDYY BY     Description
  *  -------------------------------------------------------------------
- *  010503 MF   Initial Coding
- *  031703 MF   Fix #701060. Correct logic when using a range with
- *              startIdx/endIdx. Thanks to Chris for reporting this.
- *  052603 MF   Adapt code to compile with .NET Managed C++
+ *  010503 MF     Initial Coding
+ *  031703 MF     Fix #701060. Correct logic when using a range with
+ *                startIdx/endIdx. Thanks to Chris for reporting this.
+ *  052603 MF     Adapt code to compile with .NET Managed C++
+ *  071226 MF,CC  Widen the triangular-weight factor to double: (i+1)*(i+1)
+ *                and i*(i+1) overflowed a 32-bit int at extreme periods
+ *                (past ~92682), silently returning garbage. Bit-identical
+ *                for every period where the int product fits.
  */
 
 // Import types from parent module
@@ -303,10 +308,13 @@ impl Core {
             //   the formula for a "triangular" serie is:
             //    1+2+3+3+2+1 = n*(n+1)
             //                = 3 * 4 = 12
-            // Note: entirely done with int and becomes double only
-            //       on assignement to the factor variable.
+            // Note: the (i+1) factors are widened to double so the product
+            //       cannot overflow a 32-bit int at extreme periods (i+1 reaches
+            //       ~50000 near the API maximum, and (i+1)*(i+1) exceeds INT_MAX
+            //       past period ~92682). For every period where the int product
+            //       fits, the widened value is identical.
             i = (optInTimePeriod >> 1) as usize;
-            factor = (((i + 1) * (i + 1)) as f64);
+            factor = ((i + 1) as f64) * (((i + 1)) as f64);
             factor = 1.0 / factor;
             // Initialize all the variable before
             // starting to iterate for each output.
@@ -370,7 +378,8 @@ impl Core {
             //    slightly different.
             //  - Adjustment of numeratorAdd is different. See Step (2).
             i = (optInTimePeriod >> 1) as usize;
-            factor = ((i * (i + 1)) as f64);
+            factor = (i as f64) * (((i + 1)) as f64);
+            // widen: i*(i+1) overflows int past period ~92682
             factor = 1.0 / factor;
             // Initialize all the variable before
             // starting to iterate for each output.
@@ -474,7 +483,7 @@ impl Core {
         outIdx = 0;
         if optInTimePeriod % 2 == 1 {
             i = (optInTimePeriod >> 1) as usize;
-            factor = (((i + 1) * (i + 1)) as f64);
+            factor = ((i + 1) as f64) * (((i + 1)) as f64);
             factor = 1.0 / factor;
             trailingIdx = startIdx - lookbackTotal;
             middleIdx = trailingIdx + i;
@@ -519,7 +528,7 @@ impl Core {
             }
         } else {
             i = (optInTimePeriod >> 1) as usize;
-            factor = ((i * (i + 1)) as f64);
+            factor = (i as f64) * (((i + 1)) as f64);
             factor = 1.0 / factor;
             trailingIdx = startIdx - lookbackTotal;
             middleIdx = trailingIdx + i - 1;
